+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added keyword sort option to organize tasks by keyword state Active (NOW/DOING/IN-PROGRESS), Inactive (TODO/LATER), Custom (user-defined), Waiting (WAIT/WAITING), and Completed (DONE/CANCELED/CANCELLED).
 - Added support the page properties search to filter tasks to only those that are on pages that match the property value. #26
+- Added search history for the last 10 search queries in the task list search.
 - Added styled rendering of priority indicators in the editor live preview mode, and reader view. #35
 - Fixed styling in the editor source mode for a more consistent source style that minimizes visualize restyling.